restructure fal folder
